Move test helper methods to actual codebase
